update calculate rule1

+++ app/src/main/java/com/cpeat/cpeat/data/Food.java
@@ -4,8 +4,7 @@ package com.cpeat.cpeat.data;
  * Created by CodeDiablos on 2017/7/22.
  */
 
-
 public class Food {
     public String name;
-    public Double price;
+    public float price;
 }
